# Backend-A5-Auth/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth Login & Protect API",
    "version": "1.0.0",
    "description": "A secure API using Supabase Auth. Public routes are open; protected routes require a Bearer JWT. Click Authorize, paste an access_token from /auth/login, then Try it out on the protected routes."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "example": "test@example.com" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Missing input or signup error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" }
            }
          }
        },
        "responses": {
          "200": { "description": "OK — returns access_token and refresh_token" },
          "400": { "description": "Missing input" },
          "401": {
            "description": "Invalid login credentials",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Terminate the current session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "No Content — logged out" },
          "401": { "description": "Missing/invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Public, unprotected data",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Read the authenticated user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK — user metadata" },
          "401": {
            "description": "Missing/invalid token",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "A second protected route (proves reusable middleware)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Missing/invalid token" }
        }
      }
    }
  }
}